Delete UserSettings directory
This is not needed and is causing a lot of merge commits. This directory is just for user layouts for unity and certain setting we each have set. Our settings are all different and there is no need to store this information. This directory should've been excluded from any pushes because of the .gitignore, not sure why it still made it through.